seperating tests out to run in github and jenkins seperately
