+++ gamechangerml/src/tests/feature/search/query_expansion/test_data/small_query.txt
